Load some patches earlier, instead of in our initialize method.
This is part of PloneHotfix20161129.
