+++ Products/CMFPlone/earlypatches/__init__.py
@@ -1,0 +1,2 @@
+# misc security fixes
+import security  # noqa
